Fix Carbon addDays - cast to int

+++ config/sapta.php
@@ -1,5 +1,5 @@
 <?php
 
 return [
-    'credentials_expiry_days' => env('CREDENTIALS_EXPIRY_DAYS', 7),
+    'credentials_expiry_days' => (int) env('CREDENTIALS_EXPIRY_DAYS', 7),
 ];
